requirements.txt
Flask==3.0.3
redis==5.0.8
azure-monitor-opentelemetry==1.6.4
